% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusionClust.R
\name{clusterFusions}
\alias{clusterFusions}
\title{Cluster fusions}
\usage{
clusterFusions(fusions, window = 10)
}
\arguments{
\item{fusions}{The fusions to cluster}

\item{window}{The window size to use for clustering}
}
\value{
Data frame with fusions annotated by the cluster each end belongs to
 and by paired fusion.
}
\description{
Cluster fusions
}
