% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FusionClust-package.R
\docType{package}
\name{FusionClust}
\alias{FusionClust}
\alias{FusionClust-package}
\title{FusionClust - Cluster fusions}
\description{
Collects multiple fusion points into fusion point ranges. Can be one or two
ended. Clustering is based on a window, with each fusion point initially its
own group. Groups are merged if they contain fusions that are less than or
equal to the window distance apart. For two-ended grouping, ends are
clustered separately and thenroups are merged when they contain opposite
sides of the same fusion.
}
