% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusionClust.R
\name{fusionGaps}
\alias{fusionGaps}
\title{Fusion endpoint spacings}
\usage{
fusionGaps(fusions)
}
\arguments{
\item{fusions}{Data frame of fusions to cluster}
}
\value{
A list by chromosome, each a vector of fusion endpoint spacings on
that chromosome.
}
\description{
Merges the two enpoints into a single list of positions and spacings, then
sorts, splits by chromosome, and calculates the \code{diff()} of successive
endpints.
}

