% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusionClust.R
\name{mergeCountsByCluster}
\alias{mergeCountsByCluster}
\title{Merge fusion site counts by cluster}
\usage{
mergeCountsByCluster(clusteredFusions)
}
\arguments{
\item{clusteredFusions}{The clustered fusion sites}
}
\value{
A data frame with the same columns as clusteredFusions, but with
different entries for the same cluster in the same sample collapsed.
}
\description{
Merge fusion site counts by cluster
}
