% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusionClust.R
\name{orderFusionEnds}
\alias{orderFusionEnds}
\title{Order fusion ends in the cohort.}
\usage{
orderFusionEnds(fusions)
}
\arguments{
\item{fusions}{A fusion cohort as a data frame with at least the columns:
chr1, pos1, chr2, and pos2.}
}
\value{
The input fusion cohort, but with all fusions listed low end first.
}
\description{
Cannonically ordered fusions are ordered with lower end first. This switches
them if the high end was first. Low means alphabetic ordering by chr, and
then numeric ordering by position. No fusion can have the same left and right
side as positions are the last included base on each end.
}
