% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusionClust.R
\name{readVirusCounts}
\alias{readVirusCounts}
\title{Load virus count file}
\usage{
readVirusCounts(file)
}
\arguments{
\item{file}{Single sample or cohort virus count file. Cohort has extra (first)
column giving sample name.}
}
\value{
A matrix of samples (as rownames) x virus (as columnNames), giving
the counts of viruses found in the samples.
}
\description{
Load virus count file
}
\examples{

}

