% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusionClust.R
\name{sortFusions}
\alias{sortFusions}
\title{Sort fusions}
\usage{
sortFusions(fusions)
}
\arguments{
\item{fusions}{The fusions to sort}
}
\value{
Data frame of fusion ends, sorted from low to high, with fusion
id end columns added to allow mapping back to its source.
}
\description{
Sort fusions
}

