% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusionClust.R
\name{writeAbraSv}
\alias{writeAbraSv}
\title{Write an abra.sv file}
\usage{
writeAbraSv(fusions, file, stranded = FALSE)
}
\arguments{
\item{fusions}{The file Can be a single abra sv
file, or a merged file with results for multuple samples.}

\item{file}{The abra.sv fusion data file to write.}
}
\value{
A data frame object describing the fusions, with columns:
\itemize{
  \item{ \bold{sample} [optional] The sample this fusion came from.}
  \item{ \bold{group} The group name (within sample) of equivalent fusions.}
  \item{ \bold{chr1} End1's chromosome, parsed from <chr>:<pos>.}
  \item{ \bold{pos1} End1's position, parsed from <chr>:<pos>.}
  \item{ \bold{chr1} End2's chromosome, parsed from <chr>:<pos>.}
  \item{ \bold{pos1} End2's position, parsed from <chr>:<pos>.}
  \item{ \bold{orientation} The relative directionality of end1 and end2, as
blank " " if in the same direction (FF or RR from input), or as "I" if
relatively inverted (FR or RF fromn input).}
  \item{ \bold{normal_count} The number of split reads containing the fusion
point in the normal sample.}
  \item{ \bold{tumor_count} The number of split reads containing the fusion
point in the tumor sample.}
 }
}
\description{
Write a sample's abra.sv.txt file, or a merged (concatenated)
abra.sv.txt file with an extra (initial) column giving the sample name.
}
\examples{
svFile <- "tests/testthat/data/perSampleAbraSv.tsv"
svFile <- system.file( svFile, package= "fusionClust" )
fusions <- loadAbraSv( svFile )
}

